+++ objectbox/CHANGELOG.md
@@ -7,6 +7,7 @@
   Linux, Windows). This is where the [`install.sh`](/install.sh) script downloads it by default.
   E.g. it is no longer necessary to install the library globally to run `dart test` or `flutter test`.
 * Windows: Support database directory paths that contain unicode (UTF-8) characters. #406
+* Changed `Query.stream` to collect results in a worker isolate, which should typically be faster.
 * Update: [objectbox-c 0.16.0](https://github.com/objectbox/objectbox-c/releases/tag/v0.16.0).
 * Update: [objectbox-android 3.1.3](https://github.com/objectbox/objectbox-java/releases/tag/V3.1.3).
 * Add new [task with tag list Flutter example app](example/flutter/objectbox_demo_relations) that